requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
moviepy>=1.0.3
Pillow>=9.0
requests>=2.28.0
googletrans==4.0.0-rc1
tqdm>=4.65.0
imageio-ffmpeg>=0.6.0
